// src/types/court.ts
export type Court = {
  id: number;
  name: string;
  address: string;
  location: string;
  images: string[];
  hoops: number;
  rating: number;
  bestTime: string;
  parking: boolean;
  facilities: string[];
  openingHours: string;
  restrictions: string;
  latitude: number;
  lighting: boolean;
  longitude: number;
  surface: string;
};